fix: 테이블이 생성되지 않는 문제 해결
- 하나의 DB는 DB Helper 파일을 하나만 갖도록 수정
