+++ app/src/main/java/woowacourse/shopping/database/recentlyviewedproduct/RecentlyViewedProductRepositoryImpl.kt
@@ -4,6 +4,7 @@ import android.content.ContentValues
 import android.content.Context
 import android.database.Cursor
 import woowacourse.shopping.database.ProductContract
+import woowacourse.shopping.database.ProductDBHelper
 import woowacourse.shopping.domain.Product
 import woowacourse.shopping.repository.ProductRepository
 import woowacourse.shopping.repository.RecentlyViewedProductRepository
@@ -12,7 +13,7 @@ class RecentlyViewedProductRepositoryImpl(
     context: Context,
     private val productRepository: ProductRepository,
 ) : RecentlyViewedProductRepository {
-    private val db = RecentlyViewedProductDbHelper(context).writableDatabase
+    private val db = ProductDBHelper(context).writableDatabase
 
     override fun findAll(): List<Product> {
         val products = mutableListOf<Product>()
